🗂️ types: improve type resolution
