+++ sources/@roots/bud-preset-wordpress/tsconfig.json
@@ -3,10 +3,20 @@
   "compilerOptions": {
     "rootDir": "src",
     "outDir": "lib",
+    "types": [
+      "@roots/bud-framework",
+      "@roots/bud-preset-recommend",
+      "@roots/bud-react",
+      "@roots/bud-wordpress-manifests",
+      "@roots/bud-wordpress-externals",
+      "@roots/bud-wordpress-dependencies",
+      "@roots/bud-wordpress-theme-json",
+      "@roots/bud-tailwindcss-theme-json"
+    ]
   },
+  "files": ["./src/types.ts"],
   "include": ["src"],
-  "exclude": ["./lib", "./node_modules", "**/*.test.ts", "**/*.test.tsx"],
-
+  "exclude": ["test"],
   "references": [
     {"path": "../bud/tsconfig.json"},
     {"path": "../bud-framework/tsconfig.json"},
@@ -16,6 +26,6 @@
     {"path": "../bud-wordpress-externals/tsconfig.json"},
     {"path": "../bud-wordpress-dependencies/tsconfig.json"},
     {"path": "../bud-wordpress-theme-json/tsconfig.json"},
-    {"path": "../bud-tailwindcss-theme-json/tsconfig.json"},
+    {"path": "../bud-tailwindcss-theme-json/tsconfig.json"}
   ]
 }
