+++ sources/@roots/bud-typescript/config/tsconfig.json
@@ -1,28 +1,5 @@
 {
   "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Bud",
-  "compilerOptions": {
-    "lib": [
-      "DOM",
-      "DOM.Iterable",
-      "ESNext"
-    ],
-    "module": "esnext",
-    "target": "esnext",
-    "allowJs": true,
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true,
-    "isolatedModules": true,
-    "jsx": "preserve",
-    "moduleResolution": "node",
-    "noFallthroughCasesInSwitch": true,
-    "outDir": "dist",
-    "paths": {
-      "@src/*": [
-        "src/*"
-      ]
-    },
-    "skipLibCheck": true
-  }
+  "display": "@roots/bud-typescript",
+  "extends": ["@roots/bud/config/tsconfig.json"]
 }
