+++ sources/@roots/bud/src/types.ts
@@ -1,4 +1,4 @@
-import {type Bud as BudCore} from '@roots/bud-framework/bud'
+import {type Bud as BudCore} from '@roots/bud-framework'
 
 declare module '@roots/bud' {
   interface Bud extends BudCore {}
