+++ sources/@roots/sage/src/types/index.ts
@@ -1,11 +1,6 @@
-/* eslint-disable n/no-extraneous-import */
-import '@roots/bud/types'
-import '@roots/bud-preset-wordpress/types'
-
-import type {Item} from '@roots/bud-build/item'
-import type {Loader} from '@roots/bud-build/loader'
-import type {Rule} from '@roots/bud-build/rule'
+import type {Item, Loader, Rule} from '@roots/bud-build'
 import type {PublicExtensionApi} from '@roots/bud-framework/extension'
+import type BudPresetWordPress from '@roots/bud-preset-wordpress/extension'
 
 import type Acorn from '../acorn/index.js'
 import type AcornV2PublicPath from '../acorn-v2-public-path/index.js'
@@ -27,6 +22,7 @@ declare module '@roots/bud-framework' {
     '@roots/sage/acorn': Acorn
     '@roots/sage/acorn-v2-public-path'?: AcornV2PublicPath
     '@roots/sage/blade-loader': BladeLoader
+    '@roots/bud-preset-wordpress': BudPresetWordPress
   }
 
   interface Locations {
